requirements.txt
playwright==1.48.0
pyyaml==6.0.2
click==8.1.7
